Tooltip on the canvas level
